Parity 4PCF: exact signed volume and a runtime direction grid
The parity-odd channel classified a tetrahedron's handedness from spoke
directions quantised onto a hard-wired 4x16 = 64-pixel grid, stored as one
byte per graph edge.  For tetrahedra whose three unit spoke directions are
nearly coplanar the quantisation error can flip the sign of the pixelised
triple product, or drive it to exactly zero (in which case the tetrahedron
is discarded), systematically attenuating the recovered odd amplitude.
Measured on an injection grid over five shape families, the flattened and
sheared families recovered 0.87 and 0.91 instead of 1.00, and 6.8% of all
tetrahedra were discarded as pixel-degenerate.

Two ways out, both opt-in.  The default is unchanged and bit-identical:

  -ntheta N / -nphi M   set the direction grid at run time.  The per-edge
      index is widened int8 -> int16 (N*M <= 32767).  The attenuation is
      gone by 8x32 and the discarded fraction falls to 1.4% (0.12% at
      23x91), for ~1% more query time.

  -exactparity          evaluates the signed volume from the galaxy
      positions.  No discretisation, nothing discarded, and no per-edge
      index at all: the graph reverts to 5 B/edge and 24 B/point of
      coordinates are kept instead, a net saving of about a quarter of a
      DESI-sized parity graph, for ~8% more query time.

Verified: CPU == GPU at every grid and in exact mode; the GPU out-of-core
path is bit-identical to the in-core one; mirroring a catalogue returns
-NNNN_odd configuration by configuration (exactly for the pixelized modes,
and to 1e-7 of the enumerated weight in exact mode, where machine-coplanar
tetrahedra may classify either way); validate.sh passes 4/4.

Also fixed, found while auditing the widening:

* config: the grid-limit check multiplied two default integers and could
  overflow to a value that passed the test, admitting out-of-bounds reads
  of the direction lookup table.  It now uses the division form, and
  n_dir_pixels is derived after validation rather than before.
* GPU parity: the single-pass collapse assumed 6 B/edge of resident window
  and would over-commit device memory at the wider index; it now tests the
  actual 7 B/edge against the budget it was derived from.
* graph_module used int16 without importing it, compiling only through a
  use-association leak from node_module.
* Both host-RAM estimators under-reported parity graphs.
* validate.sh and the Python wrapper now read the column header with
  comments=('#', 'r'), so outputs from older builds -- whose header wrapped
  across several physical lines -- still parse.

New tests: injection recovery versus grid resolution, cost and degeneracy
versus resolution, and the mirror-antisymmetry check.  tests/parity_shapes.py
carries the shape families and an exact port of the kernel's S4
canonicalisation so the tests can predict the sign the kernel will assign.

+++ src/node_module.F90
@@ -1,12 +1,13 @@
 module node_module
-  use iso_fortran_env, only: int32, int8
+  use iso_fortran_env, only: int32, int8, int16
   implicit none
   type :: node
     integer(kind=int32) :: nn = 0
     integer(kind=int32), allocatable :: id(:)
     integer(kind=int8), allocatable :: dist(:), mu(:)
-    ! Direction pixel index (combined theta/phi bin), allocated only for 4PCF parity
-    integer(kind=int8), allocatable :: phi(:)
+    ! Direction pixel index (combined theta/phi bin), allocated only for 4PCF
+    ! parity.  int16 so grids finer than 127 pixels are possible.
+    integer(kind=int16), allocatable :: phi(:)
   contains
     procedure :: init => node_init
     procedure :: init_with_phi => node_init_with_phi
